server/requirements.txt
fastmcp>=1.0.0
mysql-connector-python>=9.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
